porduct screen inter1

+++ src/main/java/spring/backend/m_bl5_g1_su25/OnlineShopping/ProductScreen/entity/Product.java
@@ -73,7 +73,7 @@ public class Product {
     BigDecimal salePrice;
     private String brand;
     private double rating;
-    private double discount;
+    private int discount;
     private int stock;
     private String sku;
     private int reviewCount;  // thêm
